Refactor: variables and functions
