+++ pkg/conn/connection.go
@@ -12,7 +12,7 @@ import (
 
 type Interact interface{
 	Dns() string
-	Http() string
+	Http(machineID string) string
 	Websocket() string
 	getType() string
 }
@@ -36,7 +36,7 @@ func SendDataTarget(i Interact, machineID string) string{
 	case "0":
 		return i.Dns();
 	case "1":
-		return i.Http();
+		return i.Http(machineID);
 	case "2":
 		return i.Websocket();
 	default:
@@ -47,7 +47,7 @@ func SendDataTarget(i Interact, machineID string) string{
 func (d Data) Dns() string{
 	for i:=16; i <= len(d.SecureString); i += 16{
 		chunk := d.SecureString[i-16:i];
-		ips, err := Resolv.LookupIPAddr(context.Background(),chunk+".localhost");
+		ips, err := Resolv.LookupIPAddr(context.Background(),chunk+".localhost"); // format check
 		if err != nil {
 			fmt.Println("[CONNECTION][DNS][ERROR] resolution error");
 		}
@@ -59,15 +59,14 @@ func (d Data) Dns() string{
 
 /* The connection needs to be closed, so the user can't control the destination URL to avoid SSRF */
 /* create an interface to handle the machine, to make a ID to the machine, and perform actions based on this ID, and only will work with the ID */
-func (d Data) Http() string{
-	var URL string = "http://localhost:3301/"; // set the target address
+func (d Data) Http(machineID string) string{
+	var URL string = fmt.Sprintf("http://%s:5000/",machineID) // format check
 	var command *strings.Reader = strings.NewReader(d.SecureString);
-
 	res, err := http.Post(URL, "text/html", command);
 	if err != nil {
         fmt.Println("[CONNECTION][HTTP][ERROR] invalid request");
     }
-	
+
 	body, err := ioutil.ReadAll(res.Body);
 	fmt.Println(body);
 	if err != nil {
