+++ utils/data.json
@@ -1,34 +1,1 @@
-[
-    {
-        "Id": "1",
-        "Hostname": "APOLO",
-        "Ip": "10.10.10.10",
-        "Country": "N/A",
-        "Username": "guest",
-        "Os": "WINDOWS"
-    },
-    {
-        "Id": "2",
-        "Hostname": "HERA",
-        "Ip": "10.10.10.10",
-        "Country": "N/A",
-        "Username": "guest",
-        "Os": "LINUX"
-    },
-    {
-        "Id": "3",
-        "Hostname": "HERA",
-        "Ip": "10.10.10.10",
-        "Country": "N/A",
-        "Username": "guest",
-        "Os": "LINUX"
-    },
-    {
-        "Id": "4",
-        "Hostname": "HERA",
-        "Ip": "10.10.10.10",
-        "Country": "N/A",
-        "Username": "guest",
-        "Os": "LINUX"
-    }
-]
+[{"Id":"1","Hostname":"APOLO","Ip":"10.10.10.10","Country":"N/A","Username":"guest","Os":"WINDOWS"},{"Id":"2","Hostname":"HERA","Ip":"10.10.10.10","Country":"N/A","Username":"guest","Os":"LINUX"},{"Id":"3","Hostname":"HERA","Ip":"10.10.10.10","Country":"N/A","Username":"guest","Os":"LINUX"},{"Id":"4","Hostname":"HERA","Ip":"10.10.10.10","Country":"N/A","Username":"guest","Os":"LINUX"},{"Id":"5","Hostname":"testing","Ip":"123123","Country":"x","Username":"x","Os":"x"},{"Id":"6","Hostname":"2","Ip":"2","Country":"2","Username":"2","Os":"2"}]
